Update admin page and role hierarchy
Two changes were made to improve user experience and admin functionality.

First, the navigation menu was updated to include the 'Admin' option for users with an admin role, enhancing the site's usability for admin users.

Second, a notable change was made to the role hierarchy in the 'security.yaml' file. Direct inheritance from 'ROLE_USER' to 'ROLE_ADMIN', skipping the 'ROLE_MODERATOR', simplified the structure and made admin role assignment more straightforward.

Made changes to html.twig files to fix the structure of links and changed the content of 'AdminController.php'.

Additionally, the admin page's content was scaled back to just the necessary elements for further functionality to be added in the future.