Add User Sessions functionality and interface
New UserSessionHandler class added along with required UserSessionInterface and UserSessionRepositoryInterface. This will allow the application to track user sessions. Also included are the necessary updates to the User class to add a relationship with sessions, and updates to the framework.yaml to register the new UserSessionHandler. In addition, a new migration for the users__sessions table is included. This allows for managing session data linked to user profiles, providing more granular control of user sessions.