Add "Remember Me" functionality to the login system
Added a "Remember Me" checkbox to the login form that persists user sessions for a duration of one week. This was implemented by adding a 'remember_me' field in security.yaml, adding the RememberMeBadge in the UsernamePasswordAuthenticator, and creating a corresponding checkbox in the login.ui.twig form. This change enhances the user experience, reducing the frequency of having to login for regular users.